Clean up the code a bit. Add alternate scoring.
Alternate scoring just checks to see if you searched for a specific
section or not (sort of). If any entry then gets the maximum score
(maxint), it'll just open that entry and skip the list. This could
prove to be a problem, but I haven't found one with it yet.